+++ packages/runtime-core/__tests__/scheduler.spec.ts
@@ -4,7 +4,8 @@ import {
   queuePostFlushCb,
   invalidateJob,
   queuePreFlushCb,
-  flushPreFlushCbs
+  flushPreFlushCbs,
+  flushPostFlushCbs
 } from '../src/scheduler'
 
 describe('scheduler', () => {
@@ -403,6 +404,22 @@ describe('scheduler', () => {
     expect(calls).toEqual(['job3', 'job2', 'job1'])
   })
 
+  test('sort SchedulerCbs based on id', async () => {
+    const calls: string[] = []
+    const cb1 = () => calls.push('cb1')
+    // cb1 has no id
+    const cb2 = () => calls.push('cb2')
+    cb2.id = 2
+    const cb3 = () => calls.push('cb3')
+    cb3.id = 1
+
+    queuePostFlushCb(cb1)
+    queuePostFlushCb(cb2)
+    queuePostFlushCb(cb3)
+    await nextTick()
+    expect(calls).toEqual(['cb3', 'cb2', 'cb1'])
+  })
+
   // #1595
   test('avoid duplicate postFlushCb invocation', async () => {
     const calls: string[] = []
@@ -451,7 +468,7 @@ describe('scheduler', () => {
     expect(count).toBe(1)
   })
 
-  test('should allow watcher callbacks to trigger itself', async () => {
+  test('should allow explicitly marked jobs to trigger itself', async () => {
     // normal job
     let count = 0
     const job = () => {
@@ -460,7 +477,7 @@ describe('scheduler', () => {
         queueJob(job)
       }
     }
-    job.cb = true
+    job.allowRecurse = true
     queueJob(job)
     await nextTick()
     expect(count).toBe(3)
@@ -472,7 +489,7 @@ describe('scheduler', () => {
         queuePostFlushCb(cb)
       }
     }
-    cb.cb = true
+    cb.allowRecurse = true
     queuePostFlushCb(cb)
     await nextTick()
     expect(count).toBe(5)
@@ -486,6 +503,26 @@ describe('scheduler', () => {
     job.cb = true
     queueJob(job)
     queueJob(job)
+    await nextTick()
+    expect(count).toBe(1)
+  })
+
+  // #1947 flushPostFlushCbs should handle nested calls
+  // e.g. app.mount inside app.mount
+  test('flushPostFlushCbs', async () => {
+    let count = 0
+
+    const queueAndFlush = (hook: Function) => {
+      queuePostFlushCb(hook)
+      flushPostFlushCbs()
+    }
+
+    queueAndFlush(() => {
+      queueAndFlush(() => {
+        count++
+      })
+    })
+
     await nextTick()
     expect(count).toBe(1)
   })